+++ pkg/client/notificationMock.go
@@ -7,7 +7,7 @@ import (
 	uuid "github.com/gofrs/uuid"
 )
 
-var _ NotificationV5 = (*NotificationMock)(nil)
+var _ NotificationV6 = (*NotificationMock)(nil)
 
 // NotificationMock mimics notification service >= v0.6.0  - returns information about notified users
 type NotificationMock struct {
@@ -39,7 +39,7 @@ func NewNotificationMockWithClients(upCli UserPreferences, csCli Consent) *Notif
 func (c *NotificationMock) SendTemplated(ctx context.Context,
 	templateKey, language, languageSettingKey string,
 	consentGuardKey string,
-	minConsentVersion int,
+	minConsentVersion string,
 	arbitraryEmailAddress string,
 	payload map[string]interface{},
 	subscribers ...uuid.UUID,